Project_Nine_Street/ns_requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas-ta>=0.3.14
yfinance>=0.2.30
vectorbt>=0.25.0
vollib>=0.1.5
scipy>=1.10.0
plotly>=5.15.0
pytimetk>=0.2.0
openbb>=4.1.0